requirements.txt
bluepy==1.3.0
paho-mqtt
